+++ zipper/protocols/auto/auto_group.go
@@ -141,11 +141,11 @@ type AutoGroup struct {
 	groupName string
 }
 
-func NewWithLimiter(logger *zap.Logger, config types.BackendV2, limiter limiter.ServerLimiter) (types.BackendServer, merry.Error) {
+func NewWithLimiter(logger *zap.Logger, config types.BackendV2, tldCacheDisabled bool, limiter limiter.ServerLimiter) (types.BackendServer, merry.Error) {
 	return nil, merry.New("auto group doesn't support anything useful except for New")
 }
 
-func New(logger *zap.Logger, config types.BackendV2) (types.BackendServer, merry.Error) {
+func New(logger *zap.Logger, config types.BackendV2, tldCacheDisabled bool) (types.BackendServer, merry.Error) {
 	logger = logger.With(zap.String("type", "autoGroup"), zap.String("name", config.GroupName))
 
 	limit := 100
@@ -180,7 +180,7 @@ func New(logger *zap.Logger, config types.BackendV2) (types.BackendServer, merry
 		cfg := config
 		cfg.GroupName = config.GroupName + "_" + proto
 		cfg.Servers = servers
-		c, ePtr := backendInit(logger, cfg)
+		c, ePtr := backendInit(logger, cfg, tldCacheDisabled)
 		if ePtr != nil {
 			return nil, ePtr
 		}
@@ -188,7 +188,7 @@ func New(logger *zap.Logger, config types.BackendV2) (types.BackendServer, merry
 		broadcastClients = append(broadcastClients, c)
 	}
 
-	return broadcast.NewBroadcastGroup(logger, config.GroupName+"_broadcast", broadcastClients, 600, limit, config.MaxBatchSize, *config.Timeouts)
+	return broadcast.NewBroadcastGroup(logger, config.GroupName+"_broadcast", broadcastClients, 600, limit, config.MaxBatchSize, *config.Timeouts, tldCacheDisabled)
 }
 
 func (c AutoGroup) MaxMetricsPerRequest() int {