+++ zipper/protocols/prometheus/prometheus_group.go
@@ -70,7 +70,7 @@ type PrometheusGroup struct {
 	httpQuery *helper.HttpQuery
 }
 
-func NewWithLimiter(logger *zap.Logger, config types.BackendV2, limiter limiter.ServerLimiter) (types.BackendServer, merry.Error) {
+func NewWithLimiter(logger *zap.Logger, config types.BackendV2, tldCacheDisabled bool, limiter limiter.ServerLimiter) (types.BackendServer, merry.Error) {
 	logger = logger.With(zap.String("type", "prometheus"), zap.String("protocol", config.Protocol), zap.String("name", config.GroupName))
 
 	logger.Warn("support for this backend protocol is experimental, use with caution")
@@ -159,7 +159,7 @@ func NewWithLimiter(logger *zap.Logger, config types.BackendV2, limiter limiter.
 	return c, nil
 }
 
-func New(logger *zap.Logger, config types.BackendV2) (types.BackendServer, merry.Error) {
+func New(logger *zap.Logger, config types.BackendV2, tldCacheDisabled bool) (types.BackendServer, merry.Error) {
 	if config.ConcurrencyLimit == nil {
 		return nil, types.ErrConcurrencyLimitNotSet
 	}
@@ -168,7 +168,7 @@ func New(logger *zap.Logger, config types.BackendV2) (types.BackendServer, merry
 	}
 	l := limiter.NewServerLimiter([]string{config.GroupName}, *config.ConcurrencyLimit)
 
-	return NewWithLimiter(logger, config, l)
+	return NewWithLimiter(logger, config, tldCacheDisabled, l)
 }
 
 func (c *PrometheusGroup) Children() []types.BackendServer {