+++ CHANGES.md
@@ -19,6 +19,7 @@ CHANGELOG
  - [Fix] /render and /metrics/find URLs now works correctly for format=carbonapi\_v2\_pb (protov3) and for new format
  - [Improvement] Redesign error handling and logging. Logging should be now less noisy and all error messages should contain better reasoning about error cause
  - [Improvement] Move more of the logging messages to Debug level - that should make logs less nosiy and still preserve ability to see detailed errors on Debug level
+ - [Improvement] Add a config parameter to disable tldCache (useful for clickhouse type backends)
 
 **0.12.5**
  - [Feature] Implement 'highest' function